Separate out allocators, compact and chunked
