+++ Cargo.toml
@@ -4,6 +4,18 @@ version = "0.1.0"
 authors = ["Anselm Eickhoff <anselm.eickhoff@gmail.com>"]
 license = "AGPL-3.0"
 
+[workspace]
+members = [
+    "./lib/allocators",
+    "./lib/chunked",
+    "./lib/compact",
+    "./lib/compact_macros",
+    "./lib/descartes",
+    "./lib/kay",
+    "./lib/kay_macros",
+    "./lib/monet"
+]
+
 [dependencies]
 ordered-float = "0.2.3"
 itertools = "0.5.5"
@@ -16,6 +28,12 @@ serde = "=0.8.21"
 serde_derive = "=0.8.21"
 serde_json = "=0.8.4"
 app_dirs = "1.1.1"
+
+[dependencies.compact]
+path = "./lib/compact/"
+
+[dependencies.compact_macros]
+path = "./lib/compact_macros/"
 
 [dependencies.kay]
 path = "./lib/kay/"
