+++ lib/allocators/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "allocators"
+version = "0.1.0"
+authors = ["Anselm Eickhoff <anselm.eickhoff@gmail.com>"]
+
+[dependencies]
+clippy = "0.0.106"
