+++ lib/chunked/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "chunked"
+version = "0.1.0"
+authors = ["Anselm Eickhoff <anselm.eickhoff@gmail.com>"]
+
+[dependencies]
+clippy = "0.0.106"
+
+[dependencies.allocators]
+path = "../allocators/"
