+++ lib/compact/src/compact.rs
