+++ lib/compact/src/compact_dict.rs
