+++ lib/compact/src/compact_vec.rs
