+++ lib/compact_macros/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "compact_macros"
+version = "0.1.0"
+authors = ["Anselm Eickhoff <anselm.eickhoff@gmail.com>"]
+
+[dependencies]
+syn = "=0.10.6"
+quote = "0.3"
+
+[lib]
+proc-macro = true
