+++ lib/kay/Cargo.toml
@@ -6,3 +6,12 @@ authors = ["Anselm Eickhoff <anselm.eickhoff@gmail.com>"]
 [dependencies]
 clippy = "0.0.106"
 random = "0.12.2"
+
+[dependencies.compact]
+path = "../compact/"
+
+[dependencies.compact_macros]
+path = "../compact_macros/"
+
+[dependencies.chunked]
+path = "../chunked/"
