+++ lib/monet/Cargo.toml
@@ -14,8 +14,11 @@ lazy_static = "0.2.2"
 [dependencies.kay]
 path = "../kay"
 
-[dependencies.kay_macros]
-path = "../kay_macros"
+[dependencies.compact]
+path = "../compact"
+
+[dependencies.compact_macros]
+path = "../compact_macros"
 
 [dependencies.descartes]
 path = "../descartes"
