+++ lib/monet/src/renderer/mod.rs
@@ -1,7 +1,8 @@
 
 pub use descartes::{N, P3, P2, V3, V4, M4, Iso3, Persp3, ToHomogeneous, Norm, Into2d, Into3d,
                     WithUniqueOrthogonal, Inverse, Rotate};
-use kay::{ID, Recipient, CVec, Individual, Fate};
+use compact::CVec;
+use kay::{ID, Recipient, Individual, Fate};
 
 use glium::backend::glutin_backend::GlutinFacade;
