+++ lib/monet/src/scene.rs
@@ -3,7 +3,8 @@ use std::collections::BTreeMap;
 
 pub use descartes::{N, P3, P2, V3, V4, M4, Iso3, Persp3, ToHomogeneous, Norm, Into2d, Into3d,
                     WithUniqueOrthogonal, Inverse, Rotate};
-use kay::{ID, CVec};
+use compact::CVec;
+use kay::ID;
 use fnv::FnvHashMap;
 
 use ::Batch;
