+++ src/core/geometry.rs
@@ -1,5 +1,6 @@
 use descartes::{Path, Band, Segment, P2, N, FiniteCurve, WithUniqueOrthogonal};
-use kay::{CVec, Compact, Individual};
+use compact::{CVec, Compact};
+use kay::Individual;
 use monet::{Thing, Vertex, Renderer, UpdateThing, Instance};
 
 #[derive(Compact, Clone)]
