+++ src/game/lanes_and_cars/planning/plan_result_steps.rs
@@ -1,6 +1,6 @@
 use descartes::{N, P2, Path, Norm, Band, Intersect, convex_hull, Curve, FiniteCurve,
                 RoughlyComparable, Dot, WithUniqueOrthogonal, Segment, HasBoundingBox, BoundingBox};
-use kay::{CVec, CDict};
+use compact::{CVec, CDict};
 use core::geometry::CPath;
 use core::disjoint_sets::DisjointSets;
 use ordered_float::OrderedFloat;
